PSFR-1367: removed cache 'cvl-api-get-licence-by-noms-id' as it was trying to cache a null value, also refactored code
